[ENH] Add ability to extrapolate off grids
Added `extrapolate_grid` keyword argument to InitialMassFunction.integrate_linear_piecewise_interpolated_product, which means you can extrapolate off the end of the provided grid.

If you try to extrapolate off (for example integrating from 0.1 to 100 when the grid is only defined from 1 to 25) without setting this to true (false by default), it _should_ raise an ExtrapolationError.